Databases (#34)
* typo fix - database

* deleted unnecessary classes from database - db branch

* Added functions.java
Added a new class for the database team to put all their functions. This way we don't need to edit Edb.java and can use it as reference. Cand delete or change this setup later.

* Added makeConnection class and makeConnection() constructor
This way it can be easy to pass around the connection that is needed to perform many of the querying functions

* Added createTables() to makeConnection

* Added deleteAllTables() to makeConnection

* Added populateTable(tableName, csvFileName) to makeConnection
This is similar to functions we had in the Edb.java for project B except is more efficient in that it only requires one function to read and populate any CSV file for any table.

* Added lengthFromEdges(searchType, nodeID) to makeConnection
This is to help out the algorithm team when it comes to their method of finding the shortes path to a location. More needs to be added, but it is a good start.

* Added lengthFromEdges SQL code to sqlScripts.sql

* Added L1Nodes.csv and L1Edges.csv insert statements to sqlScripts.sql

* commented out tests to prevent errors - database

* Start implementation of PreparedStatement for #5
Also:
- changed makeConnection.java's connection url path
- populated edgeLength
- changed compatibility issue with GBK encoding
- added derbytools in Gradle
Todo:
- edu/wpi/TeamE/databases/Edb.java:442 get coordinates readability

* Added Javadoc statements
Statements for makeConnection(), createTables(), deleteAllTables(), populateTable(), lengthFromEdges()

* Added !!NOTWORKING!! calculateLength trigger #5
Does not work, I am still rethinking things

* length column
Added the length to the create table hasEdge. When there is an insert on the hasEdge table, it will immediately call the addLength() which acts as a trigger and calcualtes the length between the startNode and endNode. Also added edges.csv and nodes.csv

* Cleaned code
Deleted commented out stuff that was no longer working and different debugging print statement

* JavaDoc for Algo and UI/UX fncs
#2, #3, #5, #8, #12, #13, #17, #18, #19, #29

* EASY COPY AND PAST CODE

* updated csv files (potential getEdgeInfo())

* Added deleteEdge()
deletes and edge given a nodeID

* Merged a lot of stuff
- Nodes works
.............

* Import prototype working!!!!!!!!!!!!!!!!!!!!!!

#14

* created getAllNodes()
- returns ArrayList<Node> #12

Co-authored-by: aburke921 <aburke@wpi.edu>
Co-authored-by: GP2P <73323107+GP2P@users.noreply.github.com>