Changed RequestsDB.java for #142 #172 #185
- Use assigneeID instead of assignee names
- Change empty return value of assignee from `assignee != null` to `assigneeID != 0`
- Change so that `userID >= -1` will be used to verify admin permissions instead of `userID != -1`
- Optimize imports
- Optimize sql Strings
- Change `getMyCreatedRequestInfo()` and getMyAssignedRequestInfo()` to not use `requestType` in `requests` table, but to use `Natural Join` of both tables (start to discard `requestType`)

(@Yveder Joseph  getMyCreatedRequestInfo() will have to use sql table names (floralRequests, sanitationRequest, extTransport, medDelivery, and securityServ) from now, or else you won't be able to get the assigned roonID.)

(@database we should stop keeping track of both requestType in requests because it is implementation for when we don't have tables for different requests. Let's just use from requests Natural Join medDelivery instead of an Inner Join when we use data from both so that Common attributes appear once in the result
This entity relationship is Total and Disjoint, and if we don't keep requests(moving all attributes to each request tables) we will have overlapping attributes in the relationship, and we want to avoid redundancy. For now, each request will have two records (one in requests, and one in their request type table), and they complete each other, which does not require a requestType in requests so we are not supposed to use it.)